requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
playwright==1.41.0
beautifulsoup4==4.12.3
pymongo>=4.6.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
torch>=2.0.0
